feat:  update model info type and enhance model tests
- Changed `embedding_dimension` type from `int` to `float` in `ModelInfo`.
- Improved test coverage for model listing and retrieval, ensuring compatibility
with OpenAI SDK structure.
- Added assertions to validate model attributes and metadata.
- Enhanced error handling for non-existent models.

+++ tests/llama_stack/constants.py
@@ -26,7 +26,7 @@ class ModelInfo(NamedTuple):
 
     model_id: str
     embedding_model: Model
-    embedding_dimension: int
+    embedding_dimension: float
 
 
 LLS_CORE_POD_FILTER: str = "app=llama-stack"
